# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Supply Chain AI Platform API
  description: |
    Production-grade multi-agent AI supply chain orchestration API powered by Claude Opus 4.6.

    ## Features
    - **Multi-Agent Orchestration**: 7 specialized AI agents coordinated by Claude Opus 4.6
    - **Real-time Inventory Management**: Stock tracking with automated reorder recommendations
    - **Order Processing**: End-to-end order lifecycle with fraud detection
    - **Supplier Management**: Vendor evaluation and RFQ generation
    - **Demand Forecasting**: ML-based predictions using XGBoost
    - **Anomaly Detection**: Real-time fraud and anomaly scoring

    ## Authentication
    All endpoints require JWT Bearer token authentication (except `/health`, `/ready`, and `/api/v1/auth/token`).

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    Obtain tokens via `POST /api/v1/auth/token` (OAuth2 password flow).

    ## Rate Limiting
    - 100 requests/minute per authenticated user
    - 1000 requests/hour per authenticated user
    - Agent orchestration: 10 requests/minute per user
  version: "1.0.0"
  contact:
    name: Platform Team
    email: platform@supply-chain.internal
  license:
    name: Proprietary
    url: https://supply-chain.internal/license
servers:
  - url: http://localhost:8000
    description: Local Development
  - url: https://api.supply-chain.internal
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/token:
    post:
      tags: [Authentication]
      summary: Get access token
      description: OAuth2 password flow. Returns JWT access and refresh tokens.
      security: []
      operationId: login_for_access_token
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin@supply-chain.local
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/suppliers:
    get:
      tags: [Suppliers]
      summary: List suppliers
      operationId: list_suppliers
      responses:
        "200":
          description: Supplier list

    post:
      tags: [Suppliers]
      summary: Create supplier
      operationId: create_supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierCreate"
      responses:
        "201":
          description: Supplier created
  /api/v1/suppliers/{supplier_id}/evaluate:
    post:
      tags: [Suppliers]
      summary: Evaluate supplier using AI agent
      operationId: evaluate_supplier
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Supplier evaluation complete
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 1800
...
    SupplierCreate:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        email:
          type: string
          format: email
        country:
          type: string
        lead_time_days:
          type: integer
          default: 7
        payment_terms:
          type: integer
          default: 30
...